vendor_management_system/requirements.txt
asgiref==3.7.2
Django==5.0
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
drf-yasg==1.21.7
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
packaging==23.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uuid==1.30
xlwt==1.3.0
